+++ test/jwt/override.test.js
@@ -1,0 +1,167 @@
+let assert = require("assert");
+const express = require("express");
+const request = require("supertest");
+
+const { printPath, setupST, startST, killAllST, cleanST } = require("../utils");
+let STExpress = require("../../");
+let { ProcessState } = require("../../lib/build/processState");
+let JWTRecipe = require("../../lib/build/recipe/jwt");
+
+describe(`overrideTest: ${printPath("[test/jwt/override.test.js]")}`, function () {
+    beforeEach(async function () {
+        await killAllST();
+        await setupST();
+        ProcessState.getInstance().reset();
+    });
+
+    after(async function () {
+        await killAllST();
+        await cleanST();
+    });
+
+    it("Test overriding functions", async function () {
+        await startST();
+
+        let jwtCreated = undefined;
+        let jwksKeys = undefined;
+
+        STExpress.init({
+            supertokens: {
+                connectionURI: "http://localhost:8080",
+            },
+            appInfo: {
+                apiDomain: "api.supertokens.io",
+                appName: "SuperTokens",
+                websiteDomain: "supertokens.io",
+            },
+            recipeList: [
+                JWTRecipe.init({
+                    override: {
+                        functions: (originalImplementation) => {
+                            return {
+                                ...originalImplementation,
+                                createJWT: async (input) => {
+                                    let createJWTResponse = await originalImplementation.createJWT(input);
+
+                                    if (createJWTResponse.status === "OK") {
+                                        jwtCreated = createJWTResponse.jwt;
+                                    }
+
+                                    return createJWTResponse;
+                                },
+                                getJWKS: async () => {
+                                    let getJWKSResponse = await originalImplementation.getJWKS();
+
+                                    if (getJWKSResponse.status === "OK") {
+                                        jwksKeys = getJWKSResponse.keys;
+                                    }
+
+                                    return getJWKSResponse;
+                                },
+                            };
+                        },
+                    },
+                }),
+            ],
+        });
+
+        let app = express();
+
+        app.use(STExpress.middleware());
+
+        app.use(STExpress.errorHandler());
+        app.use(express.json());
+
+        app.post("/jwtcreate", async (req, res) => {
+            let payload = req.body.payload;
+            res.json(await JWTRecipe.createJWT(payload, 1000));
+        });
+
+        let createJWTResponse = await new Promise((resolve) => {
+            request(app)
+                .post("/jwtcreate")
+                .send({
+                    payload: { someKey: "someValue" },
+                })
+                .end((err, res) => {
+                    if (err) {
+                        resolve(undefined);
+                    } else {
+                        resolve(res.body);
+                    }
+                });
+        });
+
+        assert.notStrictEqual(jwtCreated, undefined);
+        assert.deepStrictEqual(jwtCreated, createJWTResponse.jwt);
+
+        let getJWKSResponse = await new Promise((resolve) => {
+            request(app)
+                .get("/auth/jwt/jwks.json")
+                .end((err, res) => {
+                    if (err) {
+                        resolve(undefined);
+                    } else {
+                        resolve(res.body);
+                    }
+                });
+        });
+
+        assert.notStrictEqual(jwksKeys, undefined);
+        assert.deepStrictEqual(jwksKeys, getJWKSResponse.keys);
+    });
+
+    it("Test overriding APIs", async function () {
+        await startST();
+
+        let jwksKeys = undefined;
+
+        STExpress.init({
+            supertokens: {
+                connectionURI: "http://localhost:8080",
+            },
+            appInfo: {
+                apiDomain: "api.supertokens.io",
+                appName: "SuperTokens",
+                websiteDomain: "supertokens.io",
+            },
+            recipeList: [
+                JWTRecipe.init({
+                    override: {
+                        apis: (originalImplementation) => {
+                            return {
+                                ...originalImplementation,
+                                getJWKSGET: async (input) => {
+                                    let response = await originalImplementation.getJWKSGET(input);
+                                    jwksKeys = response.keys;
+                                    return response;
+                                },
+                            };
+                        },
+                    },
+                }),
+            ],
+        });
+
+        let app = express();
+
+        app.use(STExpress.middleware());
+
+        app.use(STExpress.errorHandler());
+
+        let getJWKSResponse = await new Promise((resolve) => {
+            request(app)
+                .get("/auth/jwt/jwks.json")
+                .end((err, res) => {
+                    if (err) {
+                        resolve(undefined);
+                    } else {
+                        resolve(res.body);
+                    }
+                });
+        });
+
+        assert.notStrictEqual(jwksKeys, undefined);
+        assert.deepStrictEqual(jwksKeys, getJWKSResponse.keys);
+    });
+});